refactor: Segregating context and state to an external provider
